Added update User api. Fixed Validation tests.
