+++ src/models/user.js
@@ -1,5 +1,5 @@
 const Model = require('./model.js');
-const Validation = require('../api/validation');
+const Joi = require('@hapi/joi');
 
 module.exports = class User extends Model {
 
@@ -17,44 +17,67 @@ module.exports = class User extends Model {
      * @param profilePicUrl - The new profile pic for the soon-to-be updated user
      * @param callback - Returns 'error' if the arguments did not pass the Validation check.
      */
-    static authenticate(req, profilePicUrl, callback) {
+    static validateNewUser(req, profilePicUrl, callback) {
 
-        let doValidation = false;
+        const schema = Joi.object().keys({
+            password: Joi.string().min(8).required(),
+            repeat_password: Joi.ref('password'),
+            firstName: Joi.string().pattern(/^[A-zÀ-ú\-]{2,30}$/).required(),
+            lastName: Joi.string().pattern(/^[A-zÀ-ú\-]{2,30}$/).required(),
+            primaryAddress: Joi.string().required(),
+            alternateAddress: Joi.string(),
+            email: Joi.string().email({minDomainSegments: 2, tlds: {allow: ['com', 'net']}}).required()
+        });
 
-        if (doValidation) {
-            // WARNING: Joi's validation return names are hard-coded.
-            //          Do not change the variable names "error" and "value".
-            const {error, value} = Validation.checkIfValid(req);
+        // Warning: Joi's validate return names are hard-coded -- do not alter "error" and "value" names.
+        const {error, value} = schema.validate(req.body);
 
-            if (error) {
-                callback(error);
-            } else {
-                let user = new User();
-
-                user.password = value.password;
-                user.firstName = value.firstName.charAt(0).toUpperCase() + value.firstName.slice(1);
-                user.lastName = value.lastName.charAt(0).toUpperCase() + value.lastName.slice(1);
-                user.primaryAddress = value.primaryAddress;
-                user.alternateAddress = value.alternateAddress;
-                user.imageUrl = profilePicUrl;
-                user.email = value.email;
-                callback(null, user);
-            }
-        }
-
-        else {
+        if (error) {
+            callback(error);
+        } else {
             let user = new User();
-            user.password = req.body["password:"];
-            user.firstName = req.body["firstName:"];
-            user.lastName = req.body["lastName:"];
-            user.primaryAddress = req.body["primaryAddress:"];
-            user.alternateAddress = req.body["alternateAddress:"];
+            user.password = value.password;
+            user.firstName = value.firstName;
+            user.lastName = value.lastName;
+            user.primaryAddress = value.primaryAddress;
+            user.alternateAddress = value.alternateAddress;
+            user.email = value.email;
             user.imageUrl = profilePicUrl;
-            user.email = req.body["email:"];
             callback(null, user);
         }
     }
 
+    static validateNewPassword(req, callback) {
+
+        const schema = Joi.object().keys({
+            password: Joi.string().min(8).required(),
+        });
+
+        const {error, value} = schema.validate(req.body);
+
+        if (error)
+            callback(error);
+        else
+            callback(null, value);
+    }
+
+    static validateNewDetails(req, callback) {
+
+        const schema = Joi.object().keys({
+            firstName: Joi.string().trim().pattern(/^[A-zÀ-ú\-]{2,30}$/),
+            lastName: Joi.string().trim().pattern(/^[A-zÀ-ú\-]{2,30}$/),
+            primaryAddress: Joi.string().trim(),
+            alternateAddress: Joi.string().trim(),
+        });
+
+        const {error, value} = schema.validate(req.body);
+
+        if (error)
+            callback(error);
+        else
+            callback(null, value);
+
+    }
 
     toJson() {
         return {
